requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
pyyaml>=5.4.0
tqdm>=4.62.0

# RL libraries
stable-baselines3>=1.6.0
imitation>=0.3.0
gym>=0.21.0

# Logging and visualization
wandb>=0.12.0
tensorboard>=2.8.0
matplotlib>=3.5.0

# Isaac Sim (Omniverse) - Install via Isaac Sim's Python environment
# Note: Isaac Sim should be installed separately from NVIDIA
# After installing Isaac Sim, install these packages in Isaac Sim's Python environment:
# pip install omni-isaac-gym
# pip install omni-isaac-core

# Legacy Isaac Gym Preview (optional, for backward compatibility)
# isaacgym-preview  # Only if using legacy Isaac Gym Preview
